[WIP] first steps and in modified autoplot using matplotlib. so far only 1D plots work, rest is non-functional at this point.
