+++ plottr/plot/__init__.py
@@ -1,2 +1,2 @@
-from .base import PlotNode, PlotWidgetContainer, makeFlowchartWithPlot
-from .mpl import AutoPlot as MPLAutoPlot
+from .base import PlotNode, PlotWidgetContainer, makeFlowchartWithPlot, \
+    PlotWidget
